Add files via upload
At the time of the uploading this, this is my latest project I have been working on for my Object-oriented programming (Java) class. It isn't perfect, but it works. There is a slight problem with my save button. It is supposed ot loop through the listview I created once a user types their favorite movies in the textfield and uses the add button to add it ot the listview. The save button saves it as an obvserableArrayList to th texft file "movies.txt". It is supposed to then save it as strings. Each movie having it's own line. Once you load it, it then sould load each string (movies) on a seperate line in the list view. Instead, it loads the whole array onto one line. For example, if I type "Dune, Spider-Man: Into the Spiderverse, Big Trouble in Little China, Star Wars Episode 2: Attack of the Clones" it should load as such: dune \n spider-man: into the spiderverse \n big trouble in little china \n star wars episode 2: attack of the clones. Instead it loads as such : [dune, spider-man: into the spiderverse, big trouble in little china, star wars episode 2: attack of the clones] all on one line in the list view. If anyone could help that would be much appreciated.